+++ e2e/cypress/tests/integration/files_and_attachments/upload_files_spec.js
@@ -365,4 +365,58 @@ describe('Upload Files', () => {
             });
         });
     });
+
+    it('MM-T2261 Upload SVG and post', () => {
+        const filename = 'svg.svg';
+        const aspectRatio = 1;
+
+        cy.visit(channelUrl);
+        cy.uiGetPostTextBox();
+
+        // # Attach file
+        cy.get('#advancedTextEditorCell').find('#fileUploadInput').attachFile(filename);
+        waitUntilUploadComplete();
+
+        cy.get('#create_post').find('.file-preview').within(() => {
+            // * Filename is correct
+            cy.get('.post-image__name').should('contain.text', filename);
+
+            // * Type is correct
+            cy.get('.post-image__type').should('contain.text', 'SVG');
+
+            // * Size is correct
+            cy.get('.post-image__size').should('contain.text', '6KB');
+
+            // * Img thumbnail exist
+            cy.get('.post-image__thumbnail > img').should('exist');
+        });
+
+        // # Post message
+        cy.postMessage('hello');
+        cy.uiWaitUntilMessagePostedIncludes('hello');
+
+        // # Open file preview
+        cy.uiGetFileThumbnail(filename).click();
+
+        // * Verify image preview modal is opened
+        cy.uiGetFilePreviewModal().within(() => {
+            cy.uiGetContentFilePreviewModal().find('img').should((img) => {
+                // * Image aspect ratio is maintained
+                expect(img.width() / img.height()).to.be.closeTo(aspectRatio, 1);
+            });
+
+            // * Download button should exist
+            cy.uiGetDownloadFilePreviewModal().then((downloadLink) => {
+                expect(downloadLink.attr('download')).to.equal(filename);
+
+                const fileAttachmentURL = downloadLink.attr('href');
+
+                // * Verify that download link has correct name
+                downloadAttachmentAndVerifyItsProperties(fileAttachmentURL, filename, 'attachment');
+            });
+
+            // # Close modal
+            cy.uiCloseFilePreviewModal();
+        });
+    });
 });
